+++ ios/Classes/PluginConstant.m
@@ -7,6 +7,7 @@ NSString * const onStartScan = @"onStartScan";
 NSString * const onStartConnect = @"onStartConnect";
 NSString * const onStartPrintText = @"onStartPrintText";
 NSString * const onStartPrintQrCode = @"onStartPrintQrCode";
+NSString * const onStartPrintRollLabel = @"onStartPrintRollLabel";
 NSString * const onDisconnect = @"onDisconnect";
 NSString * const isConnected = @"isConnected";
 NSString * const getAndroidSdkInt = @"getAndroidSdkInt";